Criação do TAD Minerais

+++ Minerais.h
@@ -18,16 +18,16 @@ typedef struct{
 }Minerais;
 
 void PreencheMineral(Minerais *Mineral);
-void InicializaMineral(Minerais Mineral, char Nome, double Dureza, double Reatividade, Cor Cores);
+void InicializaMineral(Minerais Mineral, char Nome, float Dureza, float Reatividade, Cor Cores);
 
 void setNomeMineral(Minerais Mineral, char Nome);
 char* getNomeMineral(Minerais *Mineral);
 
-void setDureza(Minerais *Mineral, double Dureza);
-double getDureza(Minerais *Mineral);
+void setDureza(Minerais *Mineral, float Dureza);
+float getDureza(Minerais *Mineral);
 
-void setReatividade(Minerais *Mineral, double Reatividade);
-double getReatividade(Minerais *Mineral);
+void setReatividade(Minerais *Mineral, float Reatividade);
+float getReatividade(Minerais *Mineral);
 
 void setCor(Minerais *Mineral, Cor Cores);
 Cor getCor(Minerais *Mineral);
